Configure database file for heroku
Adds a check for an environment variable called
"DATABASE_URL" which heroku uses to set up its
postgresql database. If it parses the url, it
will grab the required components and put them
in the postgresql config section. Otherwise, it
uses the variables specified in the .env file
so it can be tested locally